refactor: extract request/response types into aim_core
