+++ packages/aim_core/CHANGELOG.md
@@ -1,0 +1,3 @@
+## 0.1.1
+
+Initial release. Extracted from `aim_server` with no public API changes.
