+++ packages/aim_core/lib/src/env.dart
