+++ packages/aim_core/pubspec.yaml
@@ -1,5 +1,5 @@
-name: aim_server
-description: A lightweight, fast web framework for Dart with routing, middleware and hot reload during development.
+name: aim_core
+description: Platform-independent core of the Aim web framework. Routing, middleware, request and response types with no dart:io dependency.
 version: 0.1.1
 repository: https://github.com/aim-dart/aim
 resolution: workspace
@@ -8,7 +8,6 @@ environment:
   sdk: ^3.13.0
 
 dependencies:
-  aim_core: ^0.1.1
 
 dev_dependencies:
   lints: ^6.0.0
