+++ packages/aim_core/test/unit/request_test.dart
@@ -1,5 +1,5 @@
 import 'dart:convert';
-import 'package:aim_server/aim_server.dart';
+import 'package:aim_core/aim_core.dart';
 import 'package:test/test.dart';
 
 void main() {
