+++ pubspec.yaml
@@ -3,6 +3,7 @@ publish_to: none
 environment:
   sdk: ^3.13.0
 workspace:
+  - packages/aim_core
   - packages/aim_server
   - packages/aim_cli
   - packages/aim_server_cors
